Задача 3. Правки ревью
