+++ 03-di/01-notification-service/tasks/tasks.service.ts
@@ -1,11 +1,16 @@
 import { Injectable, NotFoundException } from "@nestjs/common";
 import { CreateTaskDto, Task, TaskStatus, UpdateTaskDto } from "./task.model";
+import { NotificationsService } from "../notifications/notifications.service";
+import { UsersService } from "../users/users.service";
 
 @Injectable()
 export class TasksService {
   private tasks: Task[] = [];
 
-  constructor() {}
+  constructor(
+    private readonly notificationsService: NotificationsService,
+    private readonly userService: UsersService,
+  ) {}
 
   async createTask(createTaskDto: CreateTaskDto) {
     const { title, description, assignedTo } = createTaskDto;
@@ -18,6 +23,13 @@ export class TasksService {
     };
     this.tasks.push(task);
 
+    const user = this.userService.getUserById(createTaskDto.assignedTo);
+    this.notificationsService.sendEmail(
+      user.email,
+      "Новая задача",
+      `Вы назначены ответственным за задачу: "${createTaskDto.title}"`,
+    );
+
     return task;
   }
 
@@ -26,8 +38,14 @@ export class TasksService {
     if (!task) {
       throw new NotFoundException(`Задача с ID ${id} не найдена`);
     }
-
     Object.assign(task, updateTaskDto);
+
+    const user = this.userService.getUserById(task.assignedTo);
+    this.notificationsService.sendSMS(
+      user.phone,
+      `Статус задачи "${task.title}" обновлён на "${task.status}"`,
+    );
+
     return task;
   }
 }
